+++ amxmodx/AMBuilder
@@ -96,6 +96,7 @@ binary.sources = [
   'stackstructs.cpp',
   'CTextParsers.cpp',
   'textparse.cpp',
+  'CvarManager.cpp',
   'cvars.cpp',
   '../public/memtools/CDetour/detours.cpp',
   '../public/memtools/CDetour/asm/asm.c',
