+++ amxmodx/CvarManager.h
