+++ amxmodx/nongpl_matches.h
@@ -25,5 +25,6 @@ struct NONGPL_CVAR_T
 
 extern NONGPL_PLUGIN_T NONGPL_PLUGIN_LIST[];
 extern NONGPL_CVAR_T NONGPL_CVAR_LIST[];
+extern bool CheckBadConList(const char *cvar, int type);
 
 #endif //_INCLUDE_AMXMODX_NONGPL_MATCHES_H_
